requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.0
bs4==0.0.1
distlib==0.3.0
filelock==3.0.12
importlib-metadata==1.5.0
lxml==4.4.1
mysql-connector==2.2.9
python-dotenv==0.12.0
pytz==2019.2
six==1.14.0
soupsieve==1.9.3
sqlparse==0.3.0
virtualenv==20.0.10
zipp==3.1.0
